+++ faultpilot/routing/__init__.py
@@ -1,0 +1,14 @@
+"""Routing domain package."""
+
+from faultpilot.routing.intent_router import IntentRouter
+from faultpilot.routing.local_classifier import LocalIntentClassifier
+from faultpilot.routing.llm_classifier import LlmIntentClassifier
+from faultpilot.routing.schemas import IntentClassification, RoutingDecision
+
+__all__ = [
+    "IntentRouter",
+    "LocalIntentClassifier",
+    "LlmIntentClassifier",
+    "IntentClassification",
+    "RoutingDecision",
+]